Fix remaining build issues, update docs
The build is somewhat more complex now than when using vanilla cbindgen
and bindgen since we now have to compile the cxxbridge separately.
Ideally we would remove this since all C++ code can be placed in the hdk
part of the plugin, and the root level CMake is only being used to
reference HDK headers and compile options.

This could perhaps be filled in by the cargo-hdk subcommand later.